+++ src/control.h
@@ -256,7 +256,7 @@ public:
   const uint8_t *cell_spikes[NUM_CELL_TYPES];
   uint32_t rast_cell_nums[NUM_CELL_TYPES];
   uint8_t **rasters[NUM_CELL_TYPES];
-  uint8_t **psths[NUM_CELL_TYPES];
+  uint32_t **psths[NUM_CELL_TYPES];
   float **pc_crs;
 
   /* voltage rasters for gui pc window */
